+++ system-parachains/asset-hubs/asset-hub-kusama/src/staking/mod.rs
@@ -27,8 +27,7 @@ use frame_support::traits::tokens::imbalance::ResolveTo;
 use pallet_election_provider_multi_block::{self as multi_block, SolutionAccuracyOf};
 use pallet_staking_async::UseValidatorsMap;
 use pallet_staking_async_rc_client as rc_client;
-use scale_info::TypeInfo;
-use sp_runtime::{transaction_validity::TransactionPriority, Perquintill};
+use sp_runtime::{generic, transaction_validity::TransactionPriority, Perquintill};
 use sp_staking::SessionIndex;
 use system_parachains_common::apis::InflationInfo;
 use xcm::v5::prelude::*;
@@ -498,7 +497,7 @@ where
 	type Extension = TxExtension;
 
 	fn create_transaction(call: RuntimeCall, extension: TxExtension) -> UncheckedExtrinsic {
-		<UncheckedExtrinsic as TypeInfo>::Identity::new_transaction(call, extension).into()
+		generic::UncheckedExtrinsic::new_transaction(call, extension).into()
 	}
 }
 
@@ -507,7 +506,7 @@ where
 	RuntimeCall: From<LocalCall>,
 {
 	fn create_bare(call: RuntimeCall) -> UncheckedExtrinsic {
-		<UncheckedExtrinsic as TypeInfo>::Identity::new_bare(call).into()
+		generic::UncheckedExtrinsic::new_bare(call).into()
 	}
 }